x-pack/plugin/sql/sql-proto/src/test/java/org/elasticsearch/xpack/sql/proto/VersionTests.java
/*
 * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
 * or more contributor license agreements. Licensed under the Elastic License;
 * you may not use this file except in compliance with the Elastic License.
 */

package org.elasticsearch.xpack.sql.proto;

import org.elasticsearch.test.ESTestCase;

public class VersionTests extends ESTestCase {
    public void test123FromString() {
        Version ver = Version.fromString("1.2.3");
        assertEquals(1, ver.major);
        assertEquals(2, ver.minor);
        assertEquals(3, ver.revision);
        assertEquals(1 * 1_000_000 + 2 * 10_000 + 3 * 100, ver.id);
        assertEquals("1.2.3", ver.version);
    }

    public void test123AlphaSnapshotFromString() {
        Version ver = Version.fromString("1.2.3-Alpha-SNAPSHOT");
        assertEquals(1, ver.major);
        assertEquals(2, ver.minor);
        assertEquals(3, ver.revision);
        assertEquals(1 * 1_000_000 + 2 * 10_000 + 3 * 100, ver.id);
        assertEquals("1.2.3-Alpha-SNAPSHOT", ver.version);
    }

    public void testVersionsEqual() {
        Version ver1 = Version.fromString("1.2.3");
        Version ver2 = Version.fromString("1.2.3");
        assertEquals(ver1, ver2);
    }

    public void testVersionsAndStringEqual() {
        Version ver1 = Version.fromString("1.2.3");
        String ver2 = "1.2.3";
        assertEquals(ver1, ver2);
    }

    public void testVersionsAndStringNotEqual() {
        Version ver1 = Version.fromString("1.2.3");
        String ver2 = "1.2.4";
        assertNotEquals(ver1, ver2);
    }

    public void testVersionsAndInvalidStringNotEqual() {
        Version ver1 = Version.fromString("1.2.3");
        String ver2 = "invalid";
        assertNotEquals(ver1, ver2);
    }

}
